+++ app/PersonioIntegration/Extensions_Base.php
@@ -14,6 +14,7 @@ use PersonioIntegrationLight\Dependencies\easySettingsForWordPress\Page;
 use PersonioIntegrationLight\Dependencies\easySettingsForWordPress\Section;
 use PersonioIntegrationLight\Dependencies\easySettingsForWordPress\Settings;
 use PersonioIntegrationLight\Dependencies\easySettingsForWordPress\Tab;
+use PersonioIntegrationLight\Helper;
 
 /**
  * Object to handle positions.
@@ -395,7 +396,7 @@ class Extensions_Base {
 		$settings_obj = Settings::get_instance();
 
 		// get hidden section.
-		$hidden = $settings_obj->get_section( 'hidden_section' );
+		$hidden = \PersonioIntegrationLight\Plugin\Settings::get_instance()->get_hidden_section();
 
 		// bail if hidden section does not exist.
 		if ( ! $hidden instanceof Section ) {
@@ -523,5 +524,14 @@ class Extensions_Base {
 
 		// get and return the extension tab.
 		return $main_settings_page->get_tab( 'extensions' );
+	}
+
+	/**
+	 * Return the URL for the form handler settings.
+	 *
+	 * @return string
+	 */
+	public function get_settings_link(): string {
+		return Helper::get_settings_url( $this->get_settings_page(), $this->get_setting_tab(), $this->get_setting_sub_tab() );
 	}
 }